HackUCI 2022 Winter Day 3 "End"
Learned to make successful graphQL queries with the PeterPortal API.
Using a list of classes (the classes the user has taken) returned by a flask/ngrok URL's arguments,
does processing to return a list of postrequisites the user is able to take.

The data returned for each postrequisite is the id, title, and description of the course.

Takes into account multiple prerequisites on a potential postrequisites, and checks al prerequisites
to check if the user has taken them.

Co-authored-by: athenarubio789@gmail.com